refactor(senpi-task): split RPC launcher resolution
